[#8] Created a basic api for user service
