+++ API/nginx.conf
@@ -3,13 +3,19 @@ events {}
 http {
     server {
         listen 80;
-        server_name backend_api;
+        server_name localhost;
 
-        location ~ ^/api/(?<endpoint>.*) {
+        location ~ ^/api/heart_rate{
             # Rewrite the URI to remove /api before passing it to the backend
-            rewrite ^/api/(.*) /$endpoint break;
-            proxy_pass http://backend_api:5000;
+            proxy_pass http://heart_rate_service:5000;
+            # Other proxy settings can be added here if needed
+        }
+
+        location ~ ^/api/users{
+            # Rewrite the URI to remove /api before passing it to the backend
+            proxy_pass http://users_service:5001;
             # Other proxy settings can be added here if needed
         }
     }
+    
 }
